+++ include/onepoint.h
@@ -1,6 +1,7 @@
 #ifndef ONEPOINT_H
 #define ONEPOINT_H
 
+#include <stdlib.h>
 #include <complex.h>
 
 #include "hmpdf.h"
@@ -22,7 +23,7 @@ onepoint_t;
 
 int null_onepoint(hmpdf_obj *d);
 int reset_onepoint(hmpdf_obj *d);
-int correct_phase1d(hmpdf_obj *d, double complex *x, int stride, int sgn);
+int correct_phase1d(hmpdf_obj *d, double complex *x, size_t stride, int sgn);
 int create_op(hmpdf_obj *d);
 int pdf_adjust_binedges(hmpdf_obj *d, int Nbins, double binedges_in[Nbins+1], double binedges_out[Nbins+1], double mean);
 int pdf_check_user_input(hmpdf_obj *d, int Nbins, double binedges[Nbins+1], int noisy);
